check

+++ scan3d/receivescan.py
@@ -1,11 +1,12 @@
 "receive"
+from pathlib import Path
 from shutil import copy2
 from compute.settings import DEVICE_PATH
-from device.proc import proc_device_data
+#from device.proc import proc_device_data
 from utils.img2img import img2img
 #from .nn_template.process import process_picture_set
 #from .nn.inference.process import proc
-from .nn.inference.config import COLOR_FILENAME, FRINGE_FILENAME, NOLIGHT_FILENAME, POINTCLOUD_JPG_FILENAME
+from .nn.inference.config import COLOR_FILENAME, FRINGE_FILENAME, NOLIGHT_FILENAME #, POINTCLOUD_JPG_FILENAME
 from .nn.inference.process_input import process_input_folder
 
 _DEBUG = False
@@ -16,6 +17,7 @@ def copy2nn(folder):
     img2img(folder / 'nolight.jpg', folder / NOLIGHT_FILENAME)
 
 def receive_scan(deviceid, folder):
+    "Receive scan from device"
     if _DEBUG:
         print(f"Receiveing scan from {deviceid} in {folder}")
     # copy file
@@ -24,4 +26,5 @@ def receive_scan(deviceid, folder):
     #proc_device_data(deviceid, folder)
     img2img(folder / 'dias.jpg', folder / FRINGE_FILENAME)
     process_input_folder(folder)
-    copy2(folder / 'pointcloud.jpg', DEVICE_PATH / deviceid / 'input' / 'last_dias.jpg' )
+    if Path.exists(folder / 'pointcloud.jpg'):
+        copy2(folder / 'pointcloud.jpg', DEVICE_PATH / deviceid / 'input' / 'last_dias.jpg' )
